fix hdrs SimTracker/TrackerMaterialAnalysis
